Finish block list UI

+++ app/src/main/java/com/jaredzhang/blockviewer/api/Request.kt
@@ -3,5 +3,5 @@ package com.jaredzhang.blockviewer.api
 import com.google.gson.annotations.SerializedName
 
 data class BlockRequest(
-    @SerializedName("block_num_or_id") private val blockNumber: Int
+    @SerializedName("block_num_or_id") private val blockNumber: Long
 )
